Trait parity_machine::Machine

pub trait Machine: for<'a> LocalizedMachine<'a> {
    type Header: Header;
    type LiveBlock: LiveBlock<Header = Self::Header>;
    type ExtendedHeader: Header;
    type EngineClient: ?Sized;
    type AuxiliaryRequest;
    type AncestryAction;
    type Error;
}

Generalization of types surrounding blockchain-suitable state machines.

Associated Types

type Header: Header

The block header type.

type LiveBlock: LiveBlock<Header = Self::Header>

The live block type.

type ExtendedHeader: Header

Block header with metadata information.

type EngineClient: ?Sized

A handle to a blockchain client for this machine.

type AuxiliaryRequest

A description of needed auxiliary data.

type AncestryAction

Actions taken on ancestry blocks when commiting a new block.

type Error

Errors which can occur when querying or interacting with the machine.